requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
filelock==3.20.2
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.0
pandas==2.3.3
python-dateutil==2.9.0.post0
python-whois==0.9.6
pytz==2025.2
requests==2.32.5
requests-file==3.0.1
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tldextract==5.3.1
tzdata==2025.3
urllib3==2.6.2
Werkzeug==3.1.4
xgboost==3.1.2
